test: initialize security tests

+++ routes/offers.js
@@ -12,6 +12,8 @@ const LIST_USER_OFFERS_VIEW = "offers/listUserOffers";
 
 const OFFERS_ENPOINT = "/offers";
 
+const USER_OFFERS_ENDPOINT = "/user/offers"; // Pathname de la vista
+
 module.exports = function (app, offersRepository) {
 
     /**
@@ -70,6 +72,9 @@ module.exports = function (app, offersRepository) {
      */
     app.get("/user/offers", async function (req, res) {
         try {
+
+            // TODO: INTEGRAR PAGINACION
+
             const offers = await offersRepository.getAllUserInSessionOffers(req.session.user);
 
             // Listado con las fechas formateadas
@@ -93,11 +98,18 @@ module.exports = function (app, offersRepository) {
             } else {
 
                 const {offerId} = req.params;
-                await offersRepository.deleteOffer(ObjectId(offerId), (isDeleted) => {
-                    if(isDeleted){
-                        res.redirect(OFFERS_ENPOINT);
-                    }
-                });
+
+                // Comprobar que la oferta pertenece al usuario en sesión
+                const offer = await offersRepository.findOffer({_id: ObjectId(offerId)}, {});
+                if (offer.seller !== req.session.user) {
+                    res.status(403).json({error: "No tienes permisos para eliminar esta oferta"});
+                } else {
+                    await offersRepository.deleteOffer(ObjectId(offerId), (isDeleted) => {
+                        if (isDeleted) {
+                            req.pathname === USER_OFFERS_ENDPOINT ? res.redirect("/offers/all") : res.redirect("/user/offers");
+                        }
+                    });
+                }
             }
 
         } catch (err) {
@@ -131,7 +143,12 @@ module.exports = function (app, offersRepository) {
                 if (i > 0 && i <= lastPage)
                     pages.push(i);
             }
-            let response = {songs: result.songs, pages: pages, currentPage: page};
+
+            let response = {
+                offers: result.offers,
+                pages: pages,
+                currentPage: page
+            };
             res.render(LIST_OFFERS_VIEW, response);
         }).catch(error => {
             res.send("Se ha producido un error al listar las ofertas " + error)
@@ -150,7 +167,10 @@ module.exports = function (app, offersRepository) {
             let check = await isBought(offer, options);
             if (!check && offer.seller.id !== user.id && offer.price <= user.wallet) {
                 user.wallet -= offer.price;
-                offersRepository.buyOffer({user: user, offerId: offerId}, function (offerId) {
+                offersRepository.buyOffer({
+                    user: user,
+                    offerId: offerId
+                }, function (offerId) {
                     if (offerId == null) {
                         user.wallet += offer.price; // deshacer operación
                         res.render(LIST_OFFERS_VIEW, {buyError: true});
@@ -204,7 +224,11 @@ module.exports = function (app, offersRepository) {
             let filter = {"_id": {$in: purchasedOffers}};
             let options = {sort: {title: 1}};
             offersRepository.getOffers(filter, options).then(offers => {
-                res.render("purchase.twig", {offers: offers, pages: pages, currentPage: page});
+                res.render("purchase.twig", {
+                    offers: offers,
+                    pages: pages,
+                    currentPage: page
+                });
             }).catch(error => {
                 res.send("Se ha producido un error al listar las ofertas compradas por el usuario: " + error);
             });
@@ -225,7 +249,7 @@ module.exports = function (app, offersRepository) {
             if (user.wallet >= 20) { // coste de destacar una oferta
                 user.wallet -= 20;
                 offer.featured = true;
-                offersRepository.featuredOffer(offer,filter, options).then(result => {
+                offersRepository.featuredOffer(offer, filter, options).then(result => {
                     if (result == null)
                         res.send("Error al destacar la oferta");
                     else
